Repo org: move high-level pages to sections/, UI elements to ui/
- Moved portfolio/* pages to sections/
- Moved InteractiveElements and ServiceCard into ui/
- Updated imports in SlideshowPortfolio and LandingPage
- Prepares repo for clearer separation: sections vs ui components
