+++ src/components/ui/InteractiveElements.tsx
